Split actuator logic into machine and validations.go
- Actuator logic has become huge and we want to split it into easier to manage files.
- Enable us to create unit tests for all machine and validations logic.
